filename change/login
I changed the filename from index2.php to something that makes more
sense for the functionality of the script. I have a line in the script
to output passwords to a text file called passwords.txt so you might
need to create that file and change the permissions before you run the
script or it might give you some errors.